requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==4.0.4
django-celery-results==2.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
eventlet==0.33.1
gevent==21.12.0
greenlet==1.1.2
kombu==5.2.4
nulltype==2.3.1
packaging==21.3
pdf2image==1.16.0
Pillow==9.1.1
plaid-python==9.6.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
